# api-docs.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Black Angel Official Store API",
        "description": "API documentation for Black Angel Official Store e-commerce platform",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@blackangel.store"
        }
    },
    "servers": [
        {
            "url": "https://api.blackangel.store",
            "description": "Production server"
        },
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/auth/signin": {
            "post": {
                "summary": "Admin Login",
                "description": "Authenticate admin user with email and password",
                "tags": [
                    "Authentication"
                ],
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@blackangel.store"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "securepassword123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - Rate limited",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RateLimitError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/products": {
            "get": {
                "summary": "List Products",
                "description": "Get paginated list of products with optional filtering",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 12
                        },
                        "description": "Items per page"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search term for product names"
                    },
                    {
                        "name": "filterActive",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "active",
                                "inactive"
                            ],
                            "default": "all"
                        },
                        "description": "Filter by product status"
                    },
                    {
                        "name": "categoryId",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Filter by category ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RateLimitError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Product",
                "description": "Create a new product",
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/products/{id}": {
            "put": {
                "summary": "Update Product",
                "description": "Update an existing product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Product",
                "description": "Delete a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "summary": "List Categories",
                "description": "Get all product categories",
                "tags": [
                    "Categories"
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "List Public Products",
                "description": "Get public products for the storefront",
                "tags": [
                    "Public"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PublicProduct"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "user123"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@blackangel.store"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMIN",
                            "USER"
                        ],
                        "example": "ADMIN"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Premium Hoodie"
                    },
                    "slug": {
                        "type": "string",
                        "example": "premium-hoodie"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-quality cotton hoodie"
                    },
                    "price": {
                        "type": "integer",
                        "example": 75000
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "image1.jpg",
                            "image2.jpg"
                        ]
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PublicProduct": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    }
                }
            },
            "ProductCreate": {
                "type": "object",
                "required": [
                    "name",
                    "slug",
                    "price",
                    "images",
                    "categoryId"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2,
                        "example": "Premium Hoodie"
                    },
                    "slug": {
                        "type": "string",
                        "example": "premium-hoodie"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-quality cotton hoodie"
                    },
                    "price": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 75000
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "image1.jpg",
                            "image2.jpg"
                        ]
                    },
                    "categoryId": {
                        "type": "integer",
                        "example": 1
                    },
                    "isActive": {
                        "type": "boolean",
                        "default": true
                    }
                }
            },
            "ProductUpdate": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 2
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categoryId": {
                        "type": "integer"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Hoodies"
                    }
                }
            },
            "PaginationMeta": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "limit": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    },
                    "totalPages": {
                        "type": "integer",
                        "example": 5
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Invalid Request"
                    },
                    "detail": {
                        "type": "string",
                        "example": "name: String must contain at least 2 character(s)"
                    }
                }
            },
            "RateLimitError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Too many authentication attempts. Please try again later."
                    },
                    "retryAfter": {
                        "type": "integer",
                        "example": 900,
                        "description": "Seconds until next attempt allowed"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication endpoints"
        },
        {
            "name": "Products",
            "description": "Product management endpoints"
        },
        {
            "name": "Categories",
            "description": "Category management endpoints"
        },
        {
            "name": "Public",
            "description": "Public storefront endpoints"
        }
    ]
}